Feat: useInApp 훅 추가, 적용
