+++ src/app/main/page.tsx
@@ -14,8 +14,8 @@ import { handleAllowNotification } from '@/service/notificationPermission';
 import { setUser } from '@sentry/nextjs';
 import { getPersonalInfo } from '@/api/apis/mypageApis';
 import { app } from '@/service/initFirebase';
-import { detectInAppBrowser } from '@/utils/detectInAppBrowser';
 import FCMDisabledPrompt from '@/components/common/FCMDisabledPrompt';
+import { useInApp } from '@/hooks/useInApp';
 
 // Meeting 타입 정의
 interface IMeetingProps {
@@ -38,11 +38,11 @@ export default function Main() {
   const [meetingList, setMeetingList] = useState<IMeetingProps[] | undefined>(
     undefined
   );
-  const [isInApp, setIsInApp] = useState(false);
   const deadlineRef = useRef<HTMLDivElement>(null);
   const popularRef = useRef<HTMLDivElement>(null);
   const token = localStorage.getItem('accessToken');
   const isVerifiedUser = localStorage.getItem('isVerifiedUser');
+  const isInApp = useInApp();
 
   const { mutate: likeMutate } = usePostWishList();
 
@@ -109,13 +109,6 @@ export default function Main() {
     if (token && !notification && app) {
       handleAllowNotification();
       localStorage.setItem('notification', 'true');
-    }
-  }, []);
-
-  useEffect(() => {
-    if (typeof window !== 'undefined') {
-      const isInAppBrowser = detectInAppBrowser(window.navigator.userAgent);
-      setIsInApp(isInAppBrowser);
     }
   }, []);
