+++ src/contexts/pwaContext.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { createContext, useContext, useEffect, useState } from 'react';
 
 const PWAContext = createContext(false);
